+++ packages/adminui-vue/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+ - `UFormRow`: to display attribute description at the bottom of the content in the element
+  with the `u-form-row__description` class if the `true` passed as value of the `description` property
 
 ### Deprecated